Moving functions around for some colocated context
